pyproject.toml
[project]
name = "ai-agents-and-wordflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
